refactor + fixes
