+++ packages/e2e/general/constants.ts
